tests: rename test helpers `_internal` -> `utils`
