+++ tests/README.md
@@ -1,5 +1,5 @@
 ```bash
-tests/_internal -> tests helpers and utilities
+tests/utils -> tests helpers and utilities
 tests/integration -> integration tests
 tests/unit -> unitest
 ```
