+++ tests/integration/frameworks/onnx/test_onnx_impl.py
@@ -7,8 +7,8 @@ from sklearn.ensemble import RandomForestClassifier
 
 from bentoml.exceptions import BentoMLException
 from bentoml.onnx import ONNXModel
-from tests._internal.frameworks.sklearn_utils import sklearn_model_data
-from tests._internal.helpers import assert_have_file_extension
+from tests.utils.frameworks.sklearn_utils import sklearn_model_data
+from tests.utils.helpers import assert_have_file_extension
 
 
 def predict_arr(
