+++ tests/integration/frameworks/test_fastai_impl.py
@@ -5,8 +5,8 @@ from fastai.learner import Learner
 from fastai.torch_core import Module
 
 from bentoml.fastai import FastAIModel
-from tests._internal.frameworks.pytorch_utils import LinearModel, test_df
-from tests._internal.helpers import assert_have_file_extension
+from tests.utils.frameworks.pytorch_utils import LinearModel, test_df
+from tests.utils.helpers import assert_have_file_extension
 
 
 def get_items(_x):
