+++ tests/integration/frameworks/test_tensorflow_impl.py
@@ -6,12 +6,12 @@ import pytest
 import tensorflow as tf
 
 import bentoml.tensorflow
-from tests._internal.frameworks.tensorflow_utils import (
+from tests.utils.frameworks.tensorflow_utils import (
     KerasSequentialModel,
     NativeModel,
     NativeRaggedModel,
 )
-from tests._internal.helpers import assert_have_file_extension
+from tests.utils.helpers import assert_have_file_extension
 
 MODEL_NAME = __name__.split(".")[-1]
 TF2 = tf.__version__.startswith("2")
