+++ tests/utils/_static/chinese.jpg
